Deprecate canvas support
It seems like it will be hard to support canvas despite it being a small subset of usage on chatgpt. The issue of knowing when to generate new hoverhints is hard to solve and could lead to unexpected behavior if the code is half finished. It's also brittle to changes in the chatgpt UI