Make padding on top match bottom for hover hints
Currently, the padding above the text in a hover hint does not match the padding below the text in a hover hint. For example, in the case that we have a hover hint with just a single line of description, it should be centered directly in the middle. This is not currently the case.