Fix menu dragging issue
There's a bug with the settings menu dragging behavior where I'm able to drag even when I'm not on a menu icon. In addition, it doesn't show the dragging symbol when I'm on top of a menu icon. Instead, it shows it when I am not on a menu icon. Please look at the code. Feel free to refactor in any way. Look to reduce technical debt and clean up the code and fix the dragging behavior. Simplicity is king.